#jira UE-100278 Fix C4018 signed/unsigned mismatch #rb Ionut.Matasaru #rnx
#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: matt.peters
#ROBOMERGE-SOURCE: CL 14397243 in //UE4/Main/...
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v742-14376270)

[CL 14401364 by matt peters in Dev-Tools-Staging branch]

+++ Engine/Source/Runtime/Core/Private/HAL/LowLevelMemTracker.cpp
@@ -2340,7 +2340,7 @@ namespace LLMPrivate
 	int32 FLLMAllocator::GetBinIndex(size_t Size) const
 	{
 		int BinIndex = 0;
-		while (BinIndex < NumBins && Bins[BinIndex].BinSize < Size)
+		while (BinIndex < NumBins && static_cast<size_t>(Bins[BinIndex].BinSize) < Size)
 		{
 			++BinIndex;
 		}
